requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# HTTP Client
requests==2.31.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Logging
python-json-logger==2.0.7

# Environment Variables
python-dotenv==1.0.0

# GeoIP
geoip2==4.7.0

# Cache (Optional)
redis==5.0.1

# AI/LLM (Optional)
openai>=1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1

user-agents==2.2.0